+++ src/components/Selector/Selector.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 
 import { EmojiMenu } from "../../data/emojiMenus";
+import useDnD from "../../hooks/useDnD";
 import * as S from "./style";
 
 type SelectorProps = {
@@ -9,36 +10,7 @@ type SelectorProps = {
 };
 
 function Selector({ data, setData }: SelectorProps) {
-  const handleDragStart = (e: React.DragEvent<HTMLDivElement>) => {
-    const target = e.target as HTMLElement;
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setData("text/plain", String(target.id));
-  };
-
-  const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-  };
-
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
-    const target = e.target as HTMLElement;
-    const draggedId = Number(e.dataTransfer.getData("text/plain"));
-    const targetId = Number(target.id);
-
-    changeOrder(draggedId, targetId);
-  };
-
-  const changeOrder = (draggedId: number, targetId: number) => {
-    const newData = [...data];
-    const draggedOption = newData.find((v) => v.id === draggedId);
-    const draggedOptionIdx = newData.findIndex((v) => v.id === draggedId);
-    const targetOptionIdx = newData.findIndex((v) => v.id === targetId);
-
-    if (draggedOption) {
-      newData.splice(draggedOptionIdx, 1);
-      newData.splice(targetOptionIdx, 0, draggedOption);
-    }
-    setData([...newData]);
-  };
+  const [handleDragStart, handleDragEnter, handleDragLeave, handleDrop] = useDnD(data, setData);
 
   return (
     <S.SelectorContainer>
@@ -49,10 +21,14 @@ function Selector({ data, setData }: SelectorProps) {
             key={item.id}
             id={String(item.id)}
             onDragStart={handleDragStart}
-            onDragOver={handleDragOver}
+            onDragEnter={handleDragEnter}
+            onDragLeave={handleDragLeave}
+            onDragOver={(e) => e.preventDefault()}
             onDrop={handleDrop}
             draggable
-          >{`${item.emoji} ${item.name}`}</S.Item>
+          >
+            {`${item.emoji} ${item.name}`}
+          </S.Item>
         ))}
       </S.ItemContainer>
       <S.Footer>
